requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Caching==2.1.0
greenlet==3.0.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
newsapi-python==0.2.6
pytz==2024.1
requests==2.26.0
six==1.16.0
tzlocal==5.2
urllib3==1.26.7
waitress==3.0.0
Werkzeug==2.0.2
zope.event==5.0
zope.interface==6.2
